+++ tests/phpunit/tests/blog-time.php
@@ -88,7 +88,7 @@ class Blog_Time_Test extends WP_UnitTestCase {
 	}
 
 	public function test_version() {
-		$this->assertEquals( '4.0', c2c_BlogTime::version() );
+		$this->assertEquals( '4.0.1', c2c_BlogTime::version() );
 	}
 
 	public function test_hooks_plugins_loaded() {