+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## _(in-progress)_
+## 4.0.1 _(2021-07-29)_
 * Fix: Ensure inline script is only output once per pageload to prevent issue when blog time appears on page multiple times, which resulted in all acting as if static
 * New: Add `reset()` to restore memoized class variables to default values
 * Change: Note compatibility through WP 5.8+